Fix crc32 slice8 bug.
Summary:
Pre-Nehalem crc32 slice8 implementation computes incorrect
values because first 32 bits should not have been bitflipped.  Escaped
my tests because my tests stupidly was accumulating results through XOR
and the error corrected itself when the test was run on an even number
of data sets.  New algorithm is correct when accumulating by adding.

Second commit to follow that allows compatibility with both algorithms
to allow migrations.  This commit is intended for public versions of the
patch.

Test Plan:
Test harness, manual tests copying ibdata file between platforms.

Reviewed By: mcallaghan

+++ mysys/my_crc32.c
@@ -91,7 +91,7 @@ void my_fast_crc32_init(my_bool cpuid_has_crc32)
 
 STATIC_INLINE uint32 my_fast_crc32_sse42(const uchar* buf, ulong len)
 {
-  uint64 crc = (uint64)(-1);
+  uint64 crc = (uint32)(-1); // this must only set low 32 bits
 
   while (len && ((uint64)buf & 7))
   {
@@ -141,7 +141,7 @@ STATIC_INLINE uint32 my_fast_crc32_slice8(const uchar* buf, ulong len)
 {
   assert(s_fast_crc_table_initialized);
 
-  uint64 crc = (uint64)(-1);
+  uint64 crc = (uint32)(-1); // this must only set low 32 bits
 
   while (len && ((uint64)buf & 7))
   {
